回家改去
回家改去

+++ RoyNet.GameServer/Player.cs
@@ -9,7 +9,7 @@ namespace RoyNet.GameServer
     public class Player
     {
         public Server GameServer { get; private set; }
-        public int UserID { get; private set; }
+        public long UserID { get; private set; }
 
         public void Send<T>(int cmd, T package)
         {
